+++ CHANGELOG.md
@@ -20,6 +20,7 @@ List all changes after the last release here (newer on top). Each change on a se
 
 ### Changed
 
+- Admin: Do not let users to set value with decimals while adjusting stock quantity if the sales unit doesn't permit.
 - Core: Block an attempt to delete a service provider that still has associated shipping or payment methods.
 - Front: improve SEO by tuning description meta tag to product, category and CMS pages
 - Importer: index product after importing it